+++ CHANGELOG.md
@@ -5,9 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+
+## 40.7.0 - 2018-10-19
 - [Feature] Add Toolbar component
 
-## 40.6.0 - 2018-10-26
+## 40.6.0 - 2018-10-16
 - [Patch] Add border radius overview story
 - [Patch] Adjust global border-radius from 8 to 5
 - [Patch] Update Travis.yml to send alerts to #oui-alerts in Slack